Initial work on bag slot assigning and blizzard sorting
Current status is that the bagbar now shows the backpack and rightclicking the container icons will show a dropdown menu that will allow you to assign types of loot to the bag or set it to be ignored when doing the Blizzard sorting.

Trying to assign types of loot will currently trigger an error because I still need to restructure the container icons to include the portrait element.